Fix issues with d3d12 textureViews (#220)
Change makes a fix to how 3D texture UAVs and RTVs were handled
that resulted in d3d12 errors.

Change also adds a new unit test to exercise the problematic pathway.
This unit test was used to repro the issue in #220 and also used to
verify the fixes in this change.

The change also makes some additional small fix ups discovered while
investigating the original issue.